requirements.txt
# Telegram
python-telegram-bot==21.6

# Groq (LLM: NLP + Whisper STT + Vision)
groq>=0.11.0

# OpenAI (clave requerida por config.py, no se usa activamente)
openai==1.30.1

# Supabase (base de datos)
supabase==2.4.3

# Encriptación de datos sensibles
cryptography==42.0.7

# Manejo de audio (voz → texto local, fallback)
pydub==0.25.1

# Procesamiento de PDFs
pdfplumber==0.11.0

# Generación de reportes PDF
reportlab==4.2.0

# Gráficos para reportes
matplotlib==3.9.0

# Procesamiento de imágenes (OCR fallback)
Pillow==10.3.0
pytesseract==0.3.10

# Variables de entorno
python-dotenv==1.0.1

# Utilidades de fechas (recurrentes)
python-dateutil==2.9.0

# Utilidades
httpx==0.27.0
aiofiles==23.2.1

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov>=5.0.0

# QA / CI
ruff>=0.4.0
